refactor: move str_ist_truthy to config
